pyproject.toml
# Poetry pyproject.toml: https://python-poetry.org/docs/pyproject/

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "nile_upgrades"
version = "0.0.1"
description = "Nile upgrades plugin"
readme = "README.md"
authors = [
  "Eric Lau <ericglau@outlook.com>"
]
license = "MIT"
# Pypi classifiers: https://pypi.org/classifiers/
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
packages = [
  { include = "nile_upgrades", from = "src" },
]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
click = "^8.0.4"

[tool.poetry.dev-dependencies]
darglint = "^1.5.8"
isort = "^5.7.0"
pyupgrade = "^2.7.4"
black = "^20.8b1"
mypy = "^0.981"
bandit = "^1.7.0"
safety = "^1.10.3"
pytest = "^7.1.3"
pylint = "^2.6.0"
pydocstyle = "^5.1.1"
pre-commit = "^2.9.3"
flake8 = "^3.9.2"
typed-ast = "^1.5.4"
anyio = "^3.6.1"
asyncclick = "^8.1.3.4"
cairo-lang = "0.10.0"
starknet-devnet = ">= 0.3.1, <0.3.4"
openzeppelin-cairo-contracts = "^0.4.0"
cairo-nile = "^0.10.0"

[tool.poetry.plugins."nile_plugins.nre"]
"deploy_proxy" = "nile_upgrades.deploy_proxy.deploy_proxy"
"upgrade_proxy" = "nile_upgrades.upgrade_proxy.upgrade_proxy"
